refactor: 파일명 변경 시 대소문자 인식 문제로 인해 파일 삭제
