--- src/constants/Agency.ts
@@ -1,9 +1,0 @@
-// 통신사
-export const AGENCY = [
-  { label: "KT", value: "kt" },
-  { label: "SKT", value: "skt" },
-  { label: "LG U+", value: "lg" },
-  { label: "알뜰폰 KT", value: "al_kt" },
-  { label: "알뜰폰 SKT", value: "al_skt" },
-  { label: "알뜰폰 LG U+", value: "al_lg" },
-];
