--- src/constants/mocks/InquiryMockData.ts
@@ -1,165 +1,0 @@
-// 문의내역 목데이터
-export const inquiryMockData = [
-  {
-    id: 181,
-    companyName: "한국대학교",
-    customerName: "김지훈",
-    serviceType: "학교이사",
-    moveDate: "2026.04.15",
-    createdAt: "2026.03.27 09:41",
-    isNew: true,
-  },
-  {
-    id: 182,
-    companyName: "서울성모병원",
-    customerName: "이수연",
-    serviceType: "병원이사",
-    moveDate: "2026.04.10",
-    createdAt: "2026.03.26 14:22",
-    isNew: true,
-  },
-  {
-    id: 183,
-    companyName: "GS리테일",
-    customerName: "박민준",
-    serviceType: "사무실이사",
-    moveDate: "2026.04.01",
-    createdAt: "2026.03.25 11:05",
-    isNew: false,
-  },
-  {
-    id: 184,
-    companyName: "현대물류",
-    customerName: "최영호",
-    serviceType: "물류창고",
-    moveDate: "2026.03.30",
-    createdAt: "2026.03.24 16:48",
-    isNew: false,
-  },
-  {
-    id: 185,
-    companyName: "연세대학교",
-    customerName: "정다은",
-    serviceType: "학교이사",
-    moveDate: "2026.03.28",
-    createdAt: "2026.03.23 10:30",
-    isNew: false,
-  },
-  {
-    id: 186,
-    companyName: "LG화학 연구소",
-    customerName: "강태웅",
-    serviceType: "실험실이사",
-    moveDate: "2026.03.25",
-    createdAt: "2026.03.22 09:12",
-    isNew: false,
-  },
-  {
-    id: 187,
-    companyName: "HSKorea",
-    customerName: "김민서",
-    serviceType: "사무실이사",
-    moveDate: "2026.03.20",
-    createdAt: "2026.03.21 15:55",
-    isNew: false,
-  },
-  {
-    id: 188,
-    companyName: "삼성SDI",
-    customerName: "윤재원",
-    serviceType: "공장이사",
-    moveDate: "2026.03.18",
-    createdAt: "2026.03.20 13:22",
-    isNew: false,
-  },
-  {
-    id: 189,
-    companyName: "마켓컬리",
-    customerName: "오지은",
-    serviceType: "물류창고",
-    moveDate: "2026.03.15",
-    createdAt: "2026.03.19 10:09",
-    isNew: false,
-  },
-  {
-    id: 190,
-    companyName: "카카오",
-    customerName: "신동현",
-    serviceType: "사무실이사",
-    moveDate: "2026.03.12",
-    createdAt: "2026.03.18 08:54",
-    isNew: false,
-  },
-  {
-    id: 191,
-    companyName: "네이버",
-    customerName: "이현우",
-    serviceType: "사무실이사",
-    moveDate: "2026.03.10",
-    createdAt: "2026.03.17 13:20",
-    isNew: false,
-  },
-  {
-    id: 192,
-    companyName: "쿠팡",
-    customerName: "박서준",
-    serviceType: "물류창고",
-    moveDate: "2026.03.08",
-    createdAt: "2026.03.16 09:33",
-    isNew: false,
-  },
-  {
-    id: 193,
-    companyName: "롯데마트",
-    customerName: "정유진",
-    serviceType: "물류창고",
-    moveDate: "2026.03.06",
-    createdAt: "2026.03.15 14:11",
-    isNew: false,
-  },
-  {
-    id: 194,
-    companyName: "CJ대한통운",
-    customerName: "김성훈",
-    serviceType: "물류창고",
-    moveDate: "2026.03.05",
-    createdAt: "2026.03.14 11:02",
-    isNew: false,
-  },
-  {
-    id: 195,
-    companyName: "포스코",
-    customerName: "이정민",
-    serviceType: "공장이사",
-    moveDate: "2026.03.03",
-    createdAt: "2026.03.13 15:27",
-    isNew: false,
-  },
-  {
-    id: 196,
-    companyName: "한화",
-    customerName: "최지훈",
-    serviceType: "공장이사",
-    moveDate: "2026.03.01",
-    createdAt: "2026.03.12 10:45",
-    isNew: false,
-  },
-  {
-    id: 197,
-    companyName: "SK하이닉스",
-    customerName: "박지훈",
-    serviceType: "공장이사",
-    moveDate: "2026.02.28",
-    createdAt: "2026.03.11 16:18",
-    isNew: false,
-  },
-  {
-    id: 198,
-    companyName: "아모레퍼시픽",
-    customerName: "김하늘",
-    serviceType: "연구소이사",
-    moveDate: "2026.02.26",
-    createdAt: "2026.03.10 09:50",
-    isNew: false,
-  },
-];
